+++ R/crew_launcher.R
@@ -55,7 +55,7 @@
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
 #' crew_session_start()
 #' router <- crew_router()
-#' router$listen()
+#' router$start()
 #' launcher <- crew_launcher_local()
 #' launcher$populate(sockets = router$sockets)
 #' launcher$launch()
@@ -109,7 +109,7 @@ crew_launcher <- function(
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
 #' crew_session_start()
 #' router <- crew_router()
-#' router$listen()
+#' router$start()
 #' launcher <- crew_launcher_local()
 #' launcher$populate(sockets = router$sockets)
 #' launcher$launch()
@@ -171,7 +171,7 @@ crew_class_launcher <- R6::R6Class(
     #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
     #' crew_session_start()
     #' router <- crew_router()
-    #' router$listen()
+    #' router$start()
     #' launcher <- crew_launcher_local()
     #' launcher$populate(sockets = router$sockets)
     #' launcher$launch()