+++ R/crew_launcher_local.R
@@ -8,7 +8,7 @@
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
 #' crew_session_start()
 #' router <- crew_router()
-#' router$listen()
+#' router$start()
 #' launcher <- crew_launcher_local()
 #' launcher$populate(sockets = router$sockets)
 #' launcher$launch()
@@ -62,7 +62,7 @@ crew_launcher_local <- function(
 #' if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
 #' crew_session_start()
 #' router <- crew_router()
-#' router$listen()
+#' router$start()
 #' launcher <- crew_launcher_local()
 #' launcher$populate(sockets = router$sockets)
 #' launcher$launch()