+++ R/crew_retry.R
@@ -11,6 +11,9 @@
 #'   number of seconds to wait between calls to `fun`.
 #' @param seconds_timeout Nonnegative numeric of length 1,
 #'   number of seconds to loop before timing out.
+#' @param max_tries Maximum number of calls to `fun` to try
+#'   before giving up.
+#' @param error Whether to throw an error on a timeout or max tries.
 #' @param message Character of length 1, optional error message
 #'   if the wait times out.
 #' @param envir Environment to evaluate `fun`.
@@ -21,6 +24,8 @@ crew_retry <- function(
   args = list(),
   seconds_interval = 1,
   seconds_timeout = 60,
+  max_tries = Inf,
+  error = TRUE,
   message = character(0),
   envir = parent.frame()
 ) {
@@ -43,10 +48,20 @@ crew_retry <- function(
     seconds_timeout,
     is.numeric(.),
     length(.) == 1L,
-    !anyNA(.), . >= 0
+    !anyNA(.),
+    . >= 0
   )
-  start <- nanonext::mclock()
+  crew_assert(
+    max_tries,
+    is.numeric(.),
+    length(.) == 1L,
+    !anyNA(.),
+    . >= 0
+  )
+  crew_assert(error, isTRUE(.) || isFALSE(.))
   milli_timeout <- 1000 * seconds_timeout
+  tries <- 0L
+  start <- nanonext::mclock()
   while (!all(do.call(what = fun, args = args, envir = envir))) {
     if ((nanonext::mclock() - start) > milli_timeout) {
       message <- paste(
@@ -55,7 +70,25 @@ crew_retry <- function(
         "seconds.",
         message
       )
-      crew_expire(message)
+      if (error) {
+        crew_expire(message)
+      } else {
+        return(invisible())
+      }
+    }
+    tries <- tries + 1L
+    if (tries >= max_tries) {
+      message <- paste(
+        "giving up after",
+        max_tries,
+        "attempts.",
+        message
+      )
+      if (error) {
+        crew_expire(message)
+      } else {
+        return(invisible())
+      }
     }
     nanonext::msleep(1000 * seconds_interval)
   }